remove planet size attribute
it should just always be 100% since otherwise it can be confusing whether to use Orbit objectSize or Planet size and you should always use Orbit objectSize as that directly determines the orbit. Something with objectSize 100% will not orbit but if you set the Planet size lower you may not realize why.